+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * `Common`:
     * `Version`:
         * `MicroUtils`: `0.4.0` -> `0.4.1`
+* `Core`:
+    * `TelegramAPIUrlsKeeper` will fix ending of host url since this version
 
 ## 0.30.6